Add function to post score
