+++ src/index.html
@@ -32,6 +32,6 @@
                 </div>
             </form>
         </div>
-    </main> 
+    </main>
 </body>
 </html>
