+++ src/postgame.js
@@ -1,0 +1,17 @@
+async function postData(url = '', data = {}) {
+  // Default options are marked with *
+  const response = await fetch(url, {
+    method: 'POST', // *GET, POST, PUT, DELETE, etc.
+    headers: {
+      'Content-Type': 'application/json',
+      // 'Content-Type': 'application/x-www-form-urlencoded',
+    },
+    body: JSON.stringify(data), // body data type must match "Content-Type" header
+  });
+  return response.json(); // parses JSON response into native JavaScript objects
+}
+
+postData('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/', { name: 'FunLoop' })
+  .then((data) => {
+    console.log(data); // JSON data parsed by `data.json()` call
+  });
